Added sectionCtn component
