+++ src/main/www/src/components/Portal/Dashboard/DashboardProjectsAndWG.tsx
@@ -1,23 +1,10 @@
-import { Typography, createStyles, makeStyles, Container, CircularProgress } from '@material-ui/core';
+import { CircularProgress } from '@material-ui/core';
 import { darkOrange } from '../../../Constants/Constants';
 import CustomCard from '../../UIComponents/CustomCard/CustomCard';
 import { useEffect } from 'react';
 import { useState } from 'react';
 import BusinessCenterIcon from '@material-ui/icons/BusinessCenter';
-
-const useStyles = makeStyles(() =>
-  createStyles({
-    main: { padding: '90px 0 0', margin: 0, maxWidth: '100%' },
-    projectAndWGCtn: {
-      maxWidth: '100%',
-      display: 'flex',
-      flexWrap: 'wrap',
-      marginTop: 40,
-      padding: 0,
-      justifyContent: 'space-between',
-    },
-  })
-);
+import SectionCtn from '../../UIComponents/CustomContainer/SectionCtn';
 
 const projectsAndWGDemoData = [
   {
@@ -99,7 +86,6 @@ interface ProjectsAndWG {
 }
 
 export default function DashboardProjectsAndWG() {
-  const classes = useStyles();
   const [projectsAndWGData, setProjectsAndWG] = useState<Array<ProjectsAndWG>>([]);
 
   useEffect(() => {
@@ -119,12 +105,8 @@ export default function DashboardProjectsAndWG() {
   ));
 
   return (
-    <Container className={classes.main} id="projects-wg">
-      <Typography variant="h4">Projects and Working Groups</Typography>
-
-      <Container className={classes.projectAndWGCtn}>
-        {projectsAndWGData.length > 0 ? renderResourcesItems : <CircularProgress />}
-      </Container>
-    </Container>
+    <SectionCtn title="Projects and Working Groups" id="projects-wg">
+      {projectsAndWGData.length > 0 ? renderResourcesItems : <CircularProgress />}
+    </SectionCtn>
   );
 }
